# src/main/java/com/example/lab2/dto/category/NewsCategoryResponseDto.java
package com.example.lab2.dto.category;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class NewsCategoryResponseDto {
    private Long id;
    private String displayName;
}