# src/main/resources/api-specs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/v1/news/{id}:
    get:
      tags:
        - news-controller
      operationId: getNewsById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/News'
    put:
      tags:
        - news-controller
      operationId: updateNews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/News'
    delete:
      tags:
        - news-controller
      operationId: deleteNews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
  /api/v1/categories/{id}:
    get:
      tags:
        - news-category-controller
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NewsCategory'
    put:
      tags:
        - news-category-controller
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsCategoryDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NewsCategory'
    delete:
      tags:
        - news-category-controller
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
  /api/v1/news:
    get:
      tags:
        - news-controller
      operationId: getAllNews
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/News'
    post:
      tags:
        - news-controller
      operationId: createNews
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/News'
  /api/v1/categories:
    get:
      tags:
        - news-category-controller
      operationId: getAllCategories
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsCategory'
    post:
      tags:
        - news-category-controller
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsCategoryDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NewsCategoryResponseDto'
  /api/v1/news/category/{id}:
    delete:
      tags:
        - news-controller
      operationId: deleteAllNewsByCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
components:
  schemas:
    NewsDto:
      required:
        - categoryId
        - content
        - date
        - title
      type: object
      properties:
        title:
          maxLength: 100
          minLength: 0
          type: string
        content:
          maxLength: 1000
          minLength: 0
          type: string
        categoryId:
          type: integer
          format: int32
        date:
          type: string
          format: date
    News:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        category:
          $ref: '#/components/schemas/NewsCategory'
        date:
          type: string
          format: date
    NewsCategory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        displayName:
          type: string
        newsList:
          type: array
          items:
            $ref: '#/components/schemas/News'
    NewsCategoryDto:
      required:
        - categoryName
      type: object
      properties:
        categoryName:
          type: string
    NewsCategoryResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        displayName:
          type: string